[SCRIPTS] Rename upload and download scripts
